Uploading a copy from the original git repo
This is something I did a while back, probably during my time in the ind. study or somewhere in that time frame. Its a multiplayer bullet hell where one person uses a guitar to fire bullet patterns, I'm going to attempt to make it single player, and using the code for note detection make a music-generated bullet hell.